+++ scripts/postbuild-packages.ts
@@ -9,15 +9,28 @@ function copyFilesToDist() {
   ];
 
   pathsToCopy.forEach((pathArr) => {
-    const sourcePath = path.join('libs/packages', ...pathArr);
-    const distPath = path.join('dist/libs/packages', ...pathArr);
+    const sourcePath = path.join(process.cwd(), 'libs/packages', ...pathArr);
+    const distPath = path.join(process.cwd(), 'dist/libs/packages', ...pathArr);
+
+    console.log(`Copying '${sourcePath.replace(process.cwd(), '')}'...`);
+
     if (fs.existsSync(sourcePath)) {
       fs.copySync(sourcePath, distPath);
       console.log(`Successfully copied ${sourcePath} to ${distPath}`);
     } else {
-      throw `File not found: ${sourcePath}`;
+      throw new Error(`File not found: ${sourcePath}`);
     }
   });
 }
 
-copyFilesToDist();
+function postbuildPackages() {
+  console.log('Running @skyux/packages postbuild step...');
+  try {
+    copyFilesToDist();
+  } catch (err) {
+    console.error('[postbuild-packages error]', err);
+    process.exit(1);
+  }
+}
+
+postbuildPackages();