+++ scripts/utils/inline-external-resources-paths.ts
@@ -1,9 +1,6 @@
-import semver from 'semver';
 import fs from 'fs-extra';
 import glob from 'glob';
 import path from 'path';
-import { runCommand } from './utils/run-command';
-import { getDistTags } from './utils/npm-utils';
 
 /**
  *
@@ -70,7 +67,7 @@ function inlineStyleUrls(contents: string): string {
  * Replaces any references to `templateUrl` and `styleUrls` with `template` and `styles`, respectively.
  * Note: this currently only affects the UMD module since that's what StackBlitz imports.
  */
-function inlineExternalResourcesPaths(projectName: string): void {
+export function inlineExternalResourcesPaths(projectName: string): void {
   const bundlePattern = path.join(
     process.cwd(),
     `dist/libs/${projectName}/bundles/*.umd.js`
@@ -100,129 +97,3 @@ function inlineExternalResourcesPaths(projectName: string): void {
     console.warn(`The UMD bundle was not found. (wanted '${bundlePattern}')`);
   }
 }
-
-// Replaces any occurrence of '0.0.0-PLACEHOLDER' with a version number.
-function replacePlaceholderTextWithVersion(filePath: string, version: string) {
-  const contents = fs.readFileSync(filePath).toString();
-  fs.writeFileSync(filePath, contents.replace(/0\.0\.0-PLACEHOLDER/g, version));
-}
-
-async function createNpmrcFile(): Promise<void> {
-  const npmFilePath = path.join(process.cwd(), '.npmrc');
-
-  await fs.ensureFile(npmFilePath);
-
-  fs.writeFileSync(
-    npmFilePath,
-    `//registry.npmjs.org/:_authToken=${process.env.NPM_TOKEN}`
-  );
-}
-
-async function buildAndPublish(): Promise<void> {
-  try {
-    const packageJson = fs.readJsonSync('package.json');
-    const newVersion = packageJson.version;
-
-    fs.removeSync('dist');
-
-    const excludeProjects = [
-      'affected',
-      'code-examples',
-      'integration',
-      'integration-e2e',
-    ];
-
-    // Build all libraries.
-    await runCommand(
-      'npx',
-      [
-        'nx',
-        'run-many',
-        '--target=build',
-        '--all',
-        '--parallel',
-        '--maxParallel=2',
-        `--exclude=${excludeProjects.join(',')}`,
-      ],
-      {
-        stdio: 'inherit',
-      }
-    );
-
-    // Run postbuild steps.
-    await runCommand(
-      'npx',
-      [
-        'nx',
-        'run-many',
-        '--target=postbuild',
-        '--all',
-        '--parallel',
-        '--maxParallel=2',
-        `--exclude=${excludeProjects.join(',')}`,
-      ],
-      {
-        stdio: 'inherit',
-      }
-    );
-
-    // Derive project names from dist directories.
-    const libsDist = path.join('dist', 'libs/');
-    const projectNames = fs.readdirSync(libsDist);
-
-    for (const projectName of projectNames) {
-      replacePlaceholderTextWithVersion(
-        path.join(libsDist, projectName, 'package.json'),
-        newVersion
-      );
-    }
-
-    replacePlaceholderTextWithVersion(
-      path.join(
-        libsDist,
-        'packages',
-        'src/schematics/migrations/migration-collection.json'
-      ),
-      newVersion
-    );
-
-    await createNpmrcFile();
-
-    const distTags = await getDistTags('@skyux/core');
-
-    const semverData = semver.parse(newVersion);
-    const isPrerelease = semverData ? semverData.prerelease.length > 0 : false;
-
-    let npmPublishTag;
-    if (isPrerelease) {
-      if (semver.gt(newVersion, distTags.next)) {
-        npmPublishTag = '--tag=next';
-      }
-    } else {
-      if (semver.gt(newVersion, distTags.latest)) {
-        npmPublishTag = '--tag=latest';
-      }
-    }
-
-    const commandArgs = ['publish', '--access=public', '--dry-run'];
-    if (npmPublishTag) {
-      commandArgs.push(npmPublishTag);
-    }
-
-    for (const projectName of projectNames) {
-      const projectRoot = path.join(process.cwd(), libsDist, projectName);
-
-      inlineExternalResourcesPaths(projectName);
-
-      await runCommand('npm', commandArgs, {
-        cwd: projectRoot,
-        stdio: 'inherit',
-      });
-    }
-  } catch (err) {
-    console.error(err);
-    process.exit(1);
-  }
-}
-
-buildAndPublish();